+++ include/ramalloc/sys/linux.h
@@ -54,7 +54,7 @@ typedef struct ramlin_barrier
 } ramlin_barrier_t;
 
 ramfail_status_t ramlin_mkbarrier(ramlin_barrier_t *barrier_arg,
-      int capacity_arg);
+      size_t capacity_arg);
 ramfail_status_t ramlin_rmbarrier(ramlin_barrier_t *barrier_arg);
 ramfail_status_t ramlin_waitonbarrier(ramlin_barrier_t *barrier_arg);
